requirements.txt
click==6.6
cornice==1.2.1
Flask==0.11
Flask-Profile==0.2
gevent==1.1.1
greenlet==0.4.9
http-parser==0.8.3
itsdangerous==0.24
Jinja2==2.8
line-profiler==1.0
MarkupSafe==0.23
memory-profiler==0.41
objgraph==3.0.0
PasteDeploy==1.5.2
psutil==4.2.0
pyprof2calltree==1.3.2
pyramid==1.7
repoze.lru==0.6
requests==2.10.0
simplejson==3.8.2
statsd-client==1.0.4
translationstring==1.3
vaurien==1.9
vaurienclient==1.1
venusian==1.0
WebOb==1.6.1
Werkzeug==0.11.10
zope.deprecation==4.1.2
zope.interface==4.1.3
redis
locustio
pyzmq
